backend/requirements.txt
flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.5
itsdangerous==2.2.0
pandas==1.4.2
matplotlib==3.5.1
python-dotenv==1.0.1
sqlalchemy==2.0.31
kubernetes==30.1.0
numpy==1.23.0